Update AdminAuthController.php -declare strict types Update CartController.php - declare strict types - fix issues Update CheckoutController.php - declare strict types - fix issues Update DashboardController.php - declare strict types Update EpayService.php - declare strict types - fix issues Update OrdersController.php - declare strict types Update PaymentServiceInterface.php - switch type to float Update PaypalService.php - declare strict types - fix issues Update ShopController.php - declare strict types Update StripeService.php - declare strict types - fix issues

+++ app/Services/Payment/PaymentServiceInterface.php
@@ -5,8 +5,8 @@ namespace App\Services\Payment;
 interface PaymentServiceInterface
 {
     /**
-     * @param int $paymentTotal
+     * @param float $paymentTotal
      * @return void
      */
-    public function processPayment(int $paymentTotal): void;
+    public function processPayment(float $paymentTotal): void;
 }
